Reordered configs, allow | in TEST_BROWSER.
The configs are now in the order in which they should be executed. The
`|` symbol is now allowed to separate parts in the `TEST_BROWSER` env
variable to satisfy jenkins.

+++ config/selenium_config.py
@@ -1,9 +1,11 @@
 import os
+import re
 import shutil
 import subprocess
 import json
 
-from selenium.webdriver.firefox.webdriver import FirefoxProfile, FirefoxBinary
+from selenium.webdriver.firefox.webdriver import FirefoxProfile,  \
+    FirefoxBinary
 from selenium.webdriver.chrome.options import Options
 import selenic
 
@@ -56,10 +58,35 @@ if not LOGS:
     caps["record-logs"] = "false"
     caps["sauce-advisor"] = "false"
 
+#
+# The order of the configs is a balancing act
+#
 
+#
+# Perform these first because they are extremely cheap to perform.
+#
 config = Config("Linux", "FIREFOX", "31")
 config = Config("Linux", "CHROME", "38")
 
+#
+# Perform these next because IE compatibility is a major
+# issue. Finding problems early pays.
+#
+config = Config("Windows 8", "INTERNETEXPLORER", "10", caps, remote=True)
+config = Config("Windows 8.1", "INTERNETEXPLORER", "11", caps, remote=True)
+
+#
+# Perform these next because OS X compatibility is an issue. Again, we
+# want to find problems early.
+#
+config = Config("OS X 10.9", "CHROME", "38", caps, remote=True)
+config = Config("OS X 10.9", "CHROME", "37", caps, remote=True)
+# wed definitely breaks on Chrome 34.
+# config = Config("OS X 10.6", "CHROME", "34", caps, remote=True)
+
+#
+# The rest is unlikely to fail if the previous tests passed.
+#
 config = Config("Windows 8.1", "CHROME", "38", caps, remote=True)
 config = Config("Windows 8.1", "CHROME", "37", caps, remote=True)
 # wed definitely breaks on Chrome 34.
@@ -70,20 +97,8 @@ config = Config("Windows 8.1", "FIREFOX", "31", caps, remote=True)
 # Previous ESR: Nope. FF24 fails. Not worth keeping up so it is gone...
 # config = Config("Windows 8.1", "FIREFOX", "24", caps, remote=True)
 
-config = Config("Windows 8", "INTERNETEXPLORER", "10", caps, remote=True)
-config = Config("Windows 8.1", "INTERNETEXPLORER", "11", caps, remote=True)
-
-config = Config("OS X 10.9", "CHROME", "38", caps, remote=True)
-config = Config("OS X 10.9", "CHROME", "37", caps, remote=True)
-# wed definitely breaks on Chrome 34.
-# config = Config("OS X 10.6", "CHROME", "34", caps, remote=True)
-
 #
 # FAILING COMBINATIONS
-#
-# No support for native events yet:
-#
-# config = Config("Windows 8.1", "FIREFOX", "29", caps, remote=True)
 #
 # Fails due to a resizing bug in Selenium:
 #
@@ -93,51 +108,50 @@ config = Config("OS X 10.9", "CHROME", "37", caps, remote=True)
 #
 # config = Config("OS X 10.6", "FIREFOX", "..", caps, remote=True)
 #
-# Just fails:
-#
-# config = Config("Windows 7", "INTERNETEXPLORER", "9",
-#                 caps, remote=True)
-#
 
 #
 # The config is obtained from the TEST_BROWSER environment variable.
 #
 browser_env = os.environ.get("TEST_BROWSER", None)
-parts = browser_env.split(",")
-CONFIG = selenic.get_config(
-    platform=parts[0] or None, browser=parts[1] or None,
-    version=parts[2] or None)
+if browser_env:
+    parts = re.split(r"[,|]", browser_env)
+    CONFIG = selenic.get_config(
+        platform=parts[0] or None, browser=parts[1] or None,
+        version=parts[2] or None)
 
-if config.browser == "CHROME":
-    CHROME_OPTIONS = Options()
+    if CONFIG.browser == "CHROME":
+        CHROME_OPTIONS = Options()
+        #
+        # This prevents getting message shown in Chrome about
+        # --ignore-certificate-errors
+        #
+        # --test-type is an **experimental** option. Reevaluate this
+        # --use.
+        #
+        CHROME_OPTIONS.add_argument("test-type")
+
+    profile = FirefoxProfile()
+    # profile.set_preference("webdriver.log.file",
+    #                        "/tmp/firefox_webdriver.log")
+    # profile.set_preference("webdriver.firefox.logfile",
+    #                         "/tmp/firefox.log")
+
     #
-    # This prevents getting message shown in Chrome about
-    # --ignore-certificate-errors
+    # This turns off the downloading prompt in FF.
     #
-    # --test-type is an **experimental** option. Reevaluate this
-    # --use.
-    #
-    CHROME_OPTIONS.add_argument("test-type")
-
-profile = FirefoxProfile()
-# profile.set_preference("webdriver.log.file", "/tmp/firefox_webdriver.log")
-# profile.set_preference("webdriver.firefox.logfile", "/tmp/firefox.log")
-
-#
-# This turns off the downloading prompt in FF.
-#
-tmp_path = "selenium_tests/tmp"
-shutil.rmtree(tmp_path, True)
-os.makedirs(tmp_path)
-profile.set_preference("browser.download.folderList", 2)
-profile.set_preference("browser.download.manager.showWhenStarting", False)
-profile.set_preference("browser.download.dir", tmp_path)
-profile.set_preference("browser.helperApps.neverAsk.saveToDisk", "text/xml")
-FIREFOX_PROFILE = profile
-
-
-def post_execution():
+    tmp_path = "selenium_tests/tmp"
     shutil.rmtree(tmp_path, True)
+    os.makedirs(tmp_path)
+    profile.set_preference("browser.download.folderList", 2)
+    profile.set_preference("browser.download.manager.showWhenStarting",
+                           False)
+    profile.set_preference("browser.download.dir", tmp_path)
+    profile.set_preference(
+        "browser.helperApps.neverAsk.saveToDisk", "text/xml")
+    FIREFOX_PROFILE = profile
+
+    def post_execution():
+        shutil.rmtree(tmp_path, True)
 
 # May be required to get native events.
 # FIREFOX_BINARY = FirefoxBinary("/home/ldd/src/firefox-24/firefox")
